Split ClientActions
