+++ chain/client/src/sync/sync_actor.rs
@@ -1,7 +1,7 @@
 use super::adapter::{SyncMessage as ClientSyncMessage, SyncShardInfo};
 use near_async::messaging::Sender;
 use near_network::types::{PeerManagerMessageRequest, StateSyncResponse};
-use near_o11y::{handler_debug_span, OpenTelemetrySpanExt, WithSpanContext};
+use near_o11y::{handler_debug_span, WithSpanContext};
 use near_performance_metrics_macros::perf;
 use near_primitives::hash::CryptoHash;
 use near_store::ShardUId;
