+++ core/async/src/examples/mod.rs
@@ -1,3 +1,5 @@
+mod actix_component;
+mod actix_component_test;
 mod async_component;
 mod async_component_test;
 mod multi_instance_test;
